+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 - Added `HeaderAnnotations` and `HeaderAnnotationsAllSourceFiles` classes.
 - Added `korp.keep_undefined_annotations` setting, to include even undefined annotations in the Korp config.
 - Added `dateformat.pre_regex` setting.
+- Added `--json-log` flag to enable JSON format for logging.
 
 ### Changed